+++ js/photos.js
@@ -1,0 +1,3 @@
+import { generatePhotos } from './data.js';
+
+export const photos = generatePhotos();
